refactor(api): git working_tree 라우트를 전용 모듈로 추출
마지막 14개 라우트(작업트리 상태·스테이징·원격 동기화)를
api/git/working_tree.py로 이동. 핸들러 본문·경로·함수명 무변경.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 63건 동일, 가림 쌍 0건.

status/git-path 핸들러가 쓰는 Pydantic 모델 2종(GitStatusResponse·
GitPathUpdateRequest)도 정의 위치 그대로 함께 옮겼다.

계획은 이 모듈이 800줄을 넘으면 staging.py 와 sync.py 로 더 가르라고 했으나
실측 332줄이라 분할하지 않는다(YAGNI).

이로써 _legacy.py 의 라우트는 0개가 되어 2,022줄에서 13줄로 줄었다 —
집계 라우터 선언만 남았다. Task 11 에서 이 파일이 소멸한다.
