+++ src/backend/api/git/__init__.py
@@ -9,7 +9,16 @@
 깨진다(Global Constraints 2 위반).
 """
 
-from . import branches, commits, github, merge, merge_requests, remotes, repositories
+from . import (
+    branches,
+    commits,
+    github,
+    merge,
+    merge_requests,
+    remotes,
+    repositories,
+    working_tree,
+)
 from ._legacy import router
 from .commits import generate_draft_commits, generate_draft_commits_for_project
 
@@ -20,6 +29,7 @@ router.include_router(merge.router)
 router.include_router(merge_requests.router)
 router.include_router(remotes.router)
 router.include_router(repositories.router)
+router.include_router(working_tree.router)
 
 __all__ = [
     "generate_draft_commits",
